completed View budget page

+++ server/db/migrate/20230606015332_create_budgets.rb
@@ -2,7 +2,7 @@ class CreateBudgets < ActiveRecord::Migration[6.1]
   def change
     create_table :budgets do |t|
       t.belongs_to :household
-      t.decimal :amount
+      t.integer :amount
       t.string :name
     end
   end
